Refactored Shops CRUD
